Reworking the code to match the updated article text
