+++ python-sqlite-sqlalchemy/project/examples/example_3/app/artists/routes.py
