+++ python-sqlite-sqlalchemy/project/examples/example_3/app/artists/templates/artists.html
