+++ python-sqlite-sqlalchemy/project/examples/example_3/app/static/js/base.js
