+++ python-sqlite-sqlalchemy/project/examples/example_3/app/templates/base.html
@@ -11,7 +11,7 @@ Chinook Database Web Server
 {% endblock %}
 
 {% block navbar %}
-<nav class="navbar navbar-expand-lg fixed-top navbar-dark" style="background-color: #563D7C;">
+<nav class="navbar navbar-expand-lg fixed-top navbar-dark" style="background-color: #361Ddc;">
     <a class="navbar-brand" href="#">Example 8</a>
     <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNavAltMarkup" aria-controls="navbarNavAltMarkup" aria-expanded="false" aria-label="Toggle navigation">
         <span class="navbar-toggler-icon"></span>
